Hw 04 auth

+++ src/services/contacts.service.ts
@@ -1,14 +1,19 @@
 import { ContactModel } from 'models/contact.schema';
 import { ContactType } from 'types/Contact.type';
+import { ContactsQueryType } from 'types/ContactsQuery.type';
 
-export const getContactsService = async () => {
-  const contacts = await ContactModel.find();
+export const getContactsService = async (owner: string, { limit, skip, favorite, page }: ContactsQueryType) => {
+  const searchQuery: { owner: string; favorite?: boolean } = { owner };
+  if (favorite !== undefined) {
+    searchQuery.favorite = favorite;
+  }
+  const contacts = await ContactModel.find(searchQuery).select({ __v: 0 }).limit(limit!).skip(skip!);
 
-  return contacts;
+  return { contacts, page, limit };
 };
 
-export const getContactByIdService = async (contactId: string) => {
-  const contacts = await ContactModel.findById(contactId);
+export const getContactByIdService = async (contactId: string, owner: string) => {
+  const contacts = await ContactModel.findOne({ _id: contactId, owner }).select({ __v: 0 });
 
   return contacts;
 };
@@ -19,14 +24,16 @@ export const addContactService = async (contact: ContactType) => {
   return newContact;
 };
 
-export const removeContactByIdService = async (contactId: string) => {
-  const removedContact = await ContactModel.findByIdAndRemove({ _id: contactId });
+export const removeContactByIdService = async (contactId: string, owner: string) => {
+  const removedContact = await ContactModel.findByIdAndRemove({ _id: contactId, owner }).select({ __v: 0 });
 
   return removedContact;
 };
 
-export const updateContactByIdService = async (contactId: string, body: Partial<ContactType>) => {
-  const updatedContact = await ContactModel.findByIdAndUpdate({ _id: contactId }, body, { new: true });
+export const updateContactByIdService = async (contactId: string, body: Partial<ContactType>, owner: string) => {
+  const updatedContact = await ContactModel.findByIdAndUpdate({ _id: contactId, owner }, body, { new: true }).select({
+    __v: 0,
+  });
 
   return updatedContact;
 };
